services/threat-engine-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
redis==5.0.8
psycopg2-binary==2.9.9
pika==1.3.2
prometheus-fastapi-instrumentator==7.0.0
